test(identity): check current-coordinate identity traversal against the semantics
The current membership path moved from a correlated predicate over the closure to
a join against a hoisted class relation, so what needs pinning is that it still
decides the same thing -- including what a boolean `EXISTS` got for free and a
join does not: member visibility, and one row per physical edge however many
members a class has.

The historical suite's TS oracle already models exactly that, and it turns out to
model the current coordinate too: the identity service builds the closure from
nodes that are not soft-deleted and assertions valid at the write instant, which
is what the model's adjacency derives when the instant is the present. So the
ledger reader, the semantics model and the fixture move to
`identity-traversal-model.ts` and both coordinates check against one model rather
than two copies of one. The historical suite's assertions are unchanged.

The fixture gains a four-member class in which exactly one member carries the
outgoing edge -- pairing a seed with every member of its class yields k rows per
seed, so a relation that lost its uniqueness would multiply that single physical
edge by four. The new suite states the multiplicity properties on their own terms
as well as through the whole-graph comparison, because a duplicate matched by a
duplicate in the model would pass the comparison.

It also covers the read instant a reused query binds. The class relation carries
the instant its member-visibility filter compares against, so a peer whose window
closes between two executions of the same query object is the sharpest form of
the #246 freshness property: nothing about the graph changes, only the clock.

The whole suite passes unchanged on `main`, so it checks the semantics rather
than the new SQL. Three deliberate breaks confirm it bites: dropping member
visibility from the relation, dropping the `COALESCE` self fallback, and joining
the closure a third time so pairs duplicate.

`identity-current-traversal-scaling.test.ts` is the #270 measurement fixture,
opt-in via `TYPEGRAPH_PERF=1`. It varies source rows and fan-out separately
because the removed cost is their product, tops out at 100k matching edges, and
measures a single-start-row hop as well -- the shape that pays for a graph-wide
relation without benefiting from the removed rescan. It is registered in
`test-postgres.sh` so its Postgres leg has a lane.
