+++ packages/typegraph/tests/backends/integration/index.ts
@@ -14,6 +14,7 @@ export type { IntegrationStore } from "./fixtures";
 export { integrationTestGraph } from "./fixtures";
 export { registerFulltextIntegrationTests } from "./fulltext";
 export { registerIdentityIntegrationTests } from "./identity";
+export { registerCurrentIdentityTraversalTests } from "./identity-current-traversal";
 export { registerHistoricalIdentityTraversalTests } from "./identity-historical-traversal";
 export { registerIdentityImportIntegrationTests } from "./identity-import";
 export { registerIdentitySeparationIntegrationTests } from "./identity-separation";
